refractor(dynamic_role): renamed to dynamic_management
